feat: ensure that game modes persists after reload
Previously we were not setting the current state of the game to our components, even though we had the values saved in memory. This should suffice to ensure that everything gets loaded properly, including images.

Note that due to how the other files are structured we have to create some hardcoded state variables before being able to load state from memory.