feat: allow for better-save editing and date handling
These are going to be used for any given gamemode. Do note that in theory our debug values do not *actually* need to be a JSON. This is just to prevent any weird behaviour. We are not doing anything fundamentally different from the other attributes.

Date casting is needed to prevent string comparison edge cases being wrong.